Split off WasherViewController extensions into separate files to appease SwiftLint file length restriction
