+++ ExpressWash/ExpressWash/WasherViewController.swift
@@ -98,7 +98,7 @@ class WasherViewController: UIViewController {
         mapView.setUserTrackingMode(.follow, animated: true, completionHandler: nil)
     }
 
-    private func updateViews() {
+    func updateViews() {
         guard isViewLoaded else { return }
 
         jobDescriptionLabel.text = "No job right now"
@@ -313,121 +313,5 @@ class WasherViewController: UIViewController {
                 editWasherVC.washerController = washerController
             }
         }
-    }
-}
-
-extension WasherViewController: MGLMapViewDelegate {
-    func mapView(_ mapView: MGLMapView, didUpdate userLocation: MGLUserLocation?) {
-        lastKnownLat = mapView.userLocation!.coordinate.latitude
-        lastKnownLon = mapView.userLocation!.coordinate.longitude
-        print("location: \(lastKnownLat) x \(lastKnownLon)")
-
-        guard let washer = UserController.shared.sessionUser.washer,
-            washer.workStatus == true,
-            activeSwitch.isOn
-        else {
-            return
-        }
-
-        var washerRep = washer.representation
-        washerRep.currentLocationLat = lastKnownLat
-        washerRep.currentLocationLon = lastKnownLon
-
-        washerController.put(washerRep: washerRep) { (error) in
-            if let error = error {
-                print("Couldn't update washer location: \(error)")
-                DispatchQueue.main.async {
-                    let alert = UIAlertController()
-                    alert.title = "Unable to update"
-                    alert.message = "An error occurred while updating your location: \(error)"
-                    alert.addAction(UIAlertAction(title: "OK", style: .default, handler: nil))
-                    self.present(alert, animated: true, completion: nil)
-                }
-            } else {
-                self.washerController.updateWasher(washer, with: washerRep)
-            }
-        }
-    }
-
-    func addAnnotation() {
-        guard let job = job else { return }
-        let annotation = MGLPointAnnotation()
-        annotation.coordinate = CLLocationCoordinate2D(latitude: job.jobLocationLat, longitude: job.jobLocationLon)
-        annotation.title = "\(job.address)"
-        if job.address2 != nil {
-            annotation.title! += ", \(job.address2!)"
-        }
-        annotation.subtitle = jobStarted ? "Tap the button when you arrive -->" : "Tap the button when you are done -->"
-        mapView.addAnnotation(annotation)
-        mapView.setCenter(annotation.coordinate, animated: true)
-    }
-
-    func mapView(_ mapView: MGLMapView, annotationCanShowCallout annotation: MGLAnnotation) -> Bool {
-        true
-    }
-
-    func mapView(_ mapView: MGLMapView, rightCalloutAccessoryViewFor annotation: MGLAnnotation) -> UIView? {
-        let button = UIButton(type: .system)
-        button.frame = CGRect(x: 0, y: 0, width: 24, height: 24)
-        let buttonName = jobStarted ? "checkmark.seal.fill" : "mappin.and.ellipse"
-        button.setImage(UIImage(systemName: buttonName), for: .normal)
-        return button
-    }
-
-    func mapView(_ mapView: MGLMapView, annotation: MGLAnnotation, calloutAccessoryControlTapped control: UIControl) {
-        arrivedCompleteTapped(self)
-    }
-}
-
-extension WasherViewController: UIImagePickerControllerDelegate, UINavigationControllerDelegate {
-
-    func imagePickerController(_ picker: UIImagePickerController,
-                               didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey: Any]) {
-        guard let job = job else {
-            print("Don't know what this image is for!")
-            return
-        }
-
-        let nextImageNeeded: ImageEndpoint = jobStarted ? .imagesJobAfter : .imagesJobBefore
-        if let image = info[UIImagePickerController.InfoKey.originalImage] as? UIImage {
-            PhotoController.shared.uploadPhoto(image,
-                                               httpMethod: "POST",
-                                               endpoint: nextImageNeeded,
-                                               theID: Int(job.jobId)) { (data, error) in
-                if let error = error {
-                    let alert = UIAlertController()
-                    alert.title = "Upload failed"
-                    alert.message = "Unable to uplaod photo: \(error)"
-                    alert.addAction(UIAlertAction(title: "OK",
-                                                  style: .default,
-                                                  handler: nil))
-                    self.present(alert, animated: true, completion: nil)
-                }
-
-                if let data = data {
-                    let decoder = JSONDecoder()
-                    do {
-                        var rep = try decoder.decode(JobRepresentation.self, from: data)
-                        if nextImageNeeded == .imagesJobBefore {
-                            rep.timeArrived = DateFormatter.nowAsISOString
-                        } else if nextImageNeeded == .imagesJobAfter {
-                            rep.timeCompleted = DateFormatter.nowAsISOString
-                        }
-                        self.jobController.editJob(jobRepresentation: rep) { _, error in
-                            if let error = error {
-                                print("Failed to update job after uploading photo: \(error)")
-                            }
-                            DispatchQueue.main.async {
-                                self.updateViews()
-                            }
-                        }
-                    } catch {
-                        print("Couldn't decode job after uploading photo: \(error)")
-                    }
-                }
-            }
-        }
-
-        self.dismiss(animated: true, completion: nil)
     }
 }
